Check new access role in test, move deleteMessage to DBTestUtil class
